// src/shared/api/generated/model/createMockRuleDtoResponseType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Smart Mock Proxy
 * API сервиса Smart Mock Proxy: mock-ответы и проксирование на стенды
 * OpenAPI spec version: 0.1.0
 */

export type CreateMockRuleDtoResponseType = typeof CreateMockRuleDtoResponseType[keyof typeof CreateMockRuleDtoResponseType];


export const CreateMockRuleDtoResponseType = {
  INLINE_JSON: 'INLINE_JSON',
  FILE: 'FILE',
} as const;